+++ test/data.cmake
@@ -17,9 +17,9 @@ odr_test_data(
 odr_test_data(
         PATH "reference-output/odr-public"
         URL "https://github.com/opendocument-app/OpenDocument.test.output.git"
-        REVISION "28402fef766b8bdafe1668acb2062655fcc726c0")
+        REVISION "d7d7c115b18256ed4ede8a022c0551275dadb461")
 
 odr_test_data(
         PATH "reference-output/odr-private"
         URL "https://github.com/opendocument-app/OpenDocument.test-private.output.git"
-        REVISION "92bbd1d71c0e246edc4517d4b4e09533273f1878")
+        REVISION "51aa4659104e77cf6edf0ee6179718c8cf342469")